+++ Sources/KeyboardKit/_Deprecated/KeyboardContext+Deprecated.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public extension KeyboardContext {
+
+    @available(*, deprecated, message: "This is no longer used")
+    static var tempIsPreviewMode: Bool = false
+}